isolate json building code
